+++ example/lib/main.dart
@@ -78,24 +78,21 @@ class _HomePageState extends State<HomePage> {
           ElevatedText(
               text: 'requestIgnoreBatteryOptimizations',
               onPressed: () async {
-                final result =
-                    await recording.requestIgnoreBatteryOptimizations();
+                final result = await recording.requestIgnoreBatteryOptimizations();
                 text = "requestIgnoreBatteryOptimizations : $result";
                 setState(() {});
               }),
         ElevatedText(
             text: 'initialize(FlAudioSource.capture)',
             onPressed: () async {
-              final result =
-                  await recording.initialize(source: FlAudioSource.capture);
+              final result = await recording.initialize(source: FlAudioSource.capture);
               text = "initialize(FlAudioSource.capture) : $result";
               setState(() {});
             }),
         ElevatedText(
             text: 'initialize(FlAudioSource.microphone)',
             onPressed: () async {
-              final result =
-                  await recording.initialize(source: FlAudioSource.microphone);
+              final result = await recording.initialize(source: FlAudioSource.microphone);
               text = "initialize(FlAudioSource.microphone) : $result";
               setState(() {});
             }),
@@ -141,10 +138,10 @@ class _HomePageState extends State<HomePage> {
     player.startPlayer(
         codec: Codec.pcm16,
         fromDataBuffer: Uint8List.fromList(byte),
-        whenFinished: () {
+        whenFinished: () async {
           showToast('播放完毕');
-          player.stopPlayer();
-          player.dispositionStream();
+          await player.stopPlayer();
+          player.closePlayer();
         });
   }
 }
@@ -156,6 +153,5 @@ class ElevatedText extends StatelessWidget {
   final VoidCallback onPressed;
 
   @override
-  Widget build(BuildContext context) =>
-      ElevatedButton(onPressed: onPressed, child: Text(text));
+  Widget build(BuildContext context) => ElevatedButton(onPressed: onPressed, child: Text(text));
 }