+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.0
 
 * `AudioSource` changed to `FlAudioSource`
+* Support iOS microphone recording
 
 ## 0.0.1