+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.1
+
+* Fix: Unable to open audio session after recording on iOS
+
 ## 0.1.0
 
 * `AudioSource` changed to `FlAudioSource`